adding base screen, drawer and internationalization
