+++ lib/common/custom_drawer/drawer_tile.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:loja_virtual/ui/components/app_theme.dart';
-import 'package:loja_virtual/ui/screen/helpers/page_manager.dart';
-import 'package:provider/provider.dart';
 
 class DrawerTile extends StatelessWidget {
+
   final IconData iconData;
   final String title;
   final int page;
@@ -12,29 +10,23 @@ class DrawerTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-
-    final int currentPage = context.watch<PageManager>().page;
-    final Color primaryColor = Theme.of(context).primaryColor;
-
     return InkWell(
-      onTap: () {
-        context.read<PageManager>().setPage(page);
-      },
+      onTap: (){},
       child: Row(
         children: [
           Padding(
-            padding: const EdgeInsets.all(12.0),
+            padding: const EdgeInsets.all(16.0),
             child: Icon(
               iconData,
               size: 32,
-              color: currentPage == page ?  primaryColor : Colors.grey[700],
+              color: Colors.grey[700],
             ),
           ),
           Text(
             title,
             style: TextStyle(
               fontSize: 16,
-              color: currentPage == page ? primaryColor : Colors.grey[700],
+              color: Colors.grey[700]
             ),
           )
         ],
