--- lib/core/models/models.dart
@@ -1,2 +1,0 @@
-export 'user.dart';
-export 'user_manager.dart';
