--- lib/core/models/user.dart
@@ -1,7 +1,0 @@
-class User {
-
-  String email;
-  String password;
-
-  User({this.email, this.password});
-}
