--- lib/ui/components/app_button.dart
@@ -1,33 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AppButton extends StatelessWidget {
-  String text;
-  Function onPressed;
-  bool showProgress;
-
-  AppButton(this.text, {this.onPressed, this.showProgress = false});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 44,
-      child: RaisedButton(
-        color: Theme.of(context).primaryColor,
-        child: showProgress
-            ? Center(
-          child: CircularProgressIndicator(
-            valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
-          ),
-        )
-            : Text(
-          text,
-          style: TextStyle(
-            color: Colors.white,
-            fontSize: 20,
-          ),
-        ),
-        onPressed: onPressed,
-      ),
-    );
-  }
-}
