--- lib/ui/components/components.dart
@@ -1,4 +1,0 @@
-export 'app_button.dart';
-export 'app_error_message.dart';
-export 'app_text_input.dart';
-export 'app_theme.dart';
