--- lib/ui/screen/drawer/drawer.dart
@@ -1,2 +1,0 @@
-export './drawer_custom.dart';
-export './drawer_tile.dart';
