--- lib/ui/screen/helpers/page_manager.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PageManager{
-
-  final PageController _pageController;
-
-  PageManager(this._pageController);
-
-  int page = 0;
-
-  void setPage(int value){
-    if(value == page) return;
-    page = value;
-    _pageController.jumpToPage(value);
-  }
-
-}
