--- lib/ui/screen/helpers/validators.dart
@@ -1,5 +1,0 @@
-bool emailValid(String mail) {
-  final regex = RegExp(
-      r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$');
-  return regex.hasMatch(mail);
-}
